+++ src/lib/query/block.ts
@@ -23,14 +23,6 @@ export const getBlockListQueryDocument = graphql(`
   }
 `);
 
-export const getBlockCountQueryDocument = graphql(`
-  query getBlockCountQuery {
-    blocks(limit: 1, order_by: { height: desc }) {
-      height
-    }
-  }
-`);
-
 export const getBlockDetailsByHeightQueryDocument = graphql(`
   query getBlockDetailsByHeight($height: Int!) {
     blocks_by_pk(height: $height) {
@@ -45,6 +37,15 @@ export const getBlockDetailsByHeightQueryDocument = graphql(`
           }
         }
       }
+    }
+  }
+`);
+
+export const getLatestBlockInfoQueryDocument = graphql(`
+  query getLatestBlockInfo {
+    blocks(limit: 1, order_by: { height: desc }) {
+      height
+      timestamp
     }
   }
 `);